+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Next
-- Add `shake` method in `Camera`
+# next
+- new Improvements in process Tile in `TiledMap`.
+- Add `shake` method in `Camera`.
 
 # 1.4.8
 - Improvements in process Tile in `TiledMap`.