+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 1.4.9
 - new Improvements in process Tile in `TiledMap`.
 - Add `shake` method in `Camera`.